workers/worker1/requirements.txt
# ---------------- Web Server ----------------
fastapi==0.110.0
uvicorn[standard]==0.29.0

# ---------------- PyTorch (CUDA 12.1) ----------------
torch==2.2.2+cu121
torchvision==0.17.2+cu121
--extra-index-url https://download.pytorch.org/whl/cu121

# ---------------- SDXL / Diffusers ----------------
diffusers==0.27.2
transformers==4.39.3
accelerate==0.28.0
safetensors==0.4.3

# ---------------- Image Handling ----------------
pillow==10.3.0

# ---------------- Google Pub/Sub ----------------
google-cloud-pubsub==2.21.1

# ---------------- Database (SQL Server) ----------------
sqlalchemy==2.0.29
pyodbc==5.1.0

# ---------------- Media Storage ----------------
cloudinary==1.39.1

# ---------------- Config & Logging ----------------
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
loguru==0.7.2
